Optimize imports. Improve code quality. Use property instead of getter method.
